fix: resolve CI failures for Hytte-rqmo
Move sortSuggestions to suggestionsUtils.ts to fix react-refresh/only-export-components
lint error — component files must not export non-component functions.
