+++ web/src/pages/Suggestions.test.tsx
@@ -2,8 +2,8 @@
 import { describe, it, expect, vi, afterEach, beforeEach } from 'vitest'
 import { render, screen, waitFor, fireEvent, within } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
-import Suggestions, { sortSuggestions } from './Suggestions'
-import { nextRunHintKey } from './suggestionsUtils'
+import Suggestions from './Suggestions'
+import { nextRunHintKey, sortSuggestions } from './suggestionsUtils'
 import enCommon from '../../public/locales/en/common.json'
 import enSuggestions from '../../public/locales/en/suggestions.json'
 import type { Suggestion } from '../components/suggestions/SuggestionCard'
